This adds some decorators:  skip, skipIf, and skipUnless
They are intended to enable people to wire in test skipping using the
same decorators that are in the python2.7 standard library, so that
migration to python2.7 doesn't need to change any test cases, just
import the decorator from the standard library instead of testools.

This has a test for the skip decorator, but could probably use some
cleverness with clone_test_with_new_id in order to generate some
tests for all three decorators.