Allow skip/skipIf/skipUnless as class decorators
In unittest/unittest2 the skip/skipIf/skipUnless decorators can be used to
decorate either individual test methods or entire TestCase classes.
However, the testtools equivalents could previously only be used to
decorate test methods. If used on a TestCase the class would be replaced
with a function, resulting in the tests either not be discovered at all or
(if a custom test loader was used) potentially an error.

This change allows the skip decorators to be used on the TestCase subclass,
for equivalent functionality with unittest.

Fixes #205

Signed-off-by: Zane Bitter <zbitter@redhat.com>