Get rid of Python 2.4 support from skip decorator
All of the python versions we support have functools.wraps(), so there is
no need to have a soft import for it any more (indeed, we use it
unconditionally elsewhere).

Also, avoid modifying the attributes of the function passed in and relying
on functools.wraps() to copy them to the function we're actually returning.
Just add the attributes to the substitute function and return it as the
user is probably expecting.

Signed-off-by: Zane Bitter <zbitter@redhat.com>