Remove the travis config
In #290 we added a github actions config to move our testing to use gha.
However, the PR trigger wasn't working on that PR (I'm thinking because
the config was coming from a fork, but I'm not sure). Since #290 has
merged now the github actions workflow is running on the merged commits
so we no longer need to rely on travis to run CI anymore. This commit
removes the travis config file so travis will no longer be triggered. It
will also serve as a test to see if the github action workflow PR
trigger will work now that the initial config has been merged.