Replace deprecated `cgi` module usage with `email`
In Python 3.11 the standard library `cgi` module was deprecated with a
planned removal set for Python 3.13. In preparation for that removal,
this commit removes the usage of this deprecated module and replaces it
with the still supported standard library `email` module which is what
the documentation points to as an alternative for how `cgi` was
previously used. This should still be compatible with all the supported
Python versions but will be more future proof and not emit any
deprecation warnings with Python 3.11 anymore.