Reduce use of deprecated `distutils` module
Per PEP 632, `distutils` is deprecated and most uses of it can be
replaced by `setuptools`.  The only use within `testtools` is to support
the deprecated `TestCommand` that's due to be removed in the next major,
but in the meantime, adding a runtime dependency on `setuptools` for
Python 3.12+ (which removed `distutils` entirely) allows it to continue
working.