trivial: Remove vendoring notes on TestProgram
We no longer care about Python < 2.7. Or Python == 2.7 for that matter.
However, we do care about the additional functionality that our version
of TestProgram exposes, none of which has been successfully upstreamed
yet. This unfortunately means duplicating a lot of TestProgram's
implementation due to https://github.com/python/cpython/issues/67049.

Reflect this reality in the code.

Signed-off-by: Stephen Finucane <stephen@that.guru>